Run ibctest suite in build workflow
This slightly reorganizes the build workflow. It used to be a single job
that ran unit tests, then built and uploaded binaries.

Now it is three separate jobs; unit tests and ibctest are run
concurrently, and then the build and upload job happens only if the
prior two jobs succeed.